requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# DashScope (Qwen) SDK
dashscope==1.17.0

# Chinese tokenization
jieba==0.42.1

# Env var management
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# CORS support is included in FastAPI

# Optional: pinyin matching (if needed)
# pypinyin==0.49.0

# Optional: vector database (if needed)
# chromadb==0.4.18

# Dev tools
pytest==7.4.3
black==23.11.0

